+++ nengo_dl/tests/test_benchmarks.py
@@ -157,8 +157,8 @@ def test_cli():
     [
         (benchmarks.cconv(128, 64, nengo.RectifiedLinear()), False, 64, 0.7, 0.85),
         (benchmarks.cconv(128, 64, nengo.LIF()), False, 64, 1.5, 1.7),
-        (benchmarks.integrator(128, 32, nengo.RectifiedLinear()), True, 64, 0.8, 1.0),
-        (benchmarks.integrator(128, 32, nengo.LIF()), True, 64, 0.6, 0.7),
+        (benchmarks.integrator(128, 32, nengo.RectifiedLinear()), True, 64, 0.75, 0.9),
+        (benchmarks.integrator(128, 32, nengo.LIF()), True, 64, 0.55, 0.7),
         (
             benchmarks.random_network(
                 64,
@@ -195,4 +195,5 @@ def test_performance(net, train, minibatch_size, min, max):
         do_profile=False,
         reps=5,
     )
-    assert min < time < max
+    assert time > min
+    assert time < max